Refactoring code, Sensor Name Change checked for validity
